+++ backend/app/controllers/api/v1/posts_controller.rb
@@ -1,11 +1,16 @@
 require 'open-uri'
 
 class Api::V1::PostsController < ApplicationController
-  before_action :authenticate, only: [:create]
+  before_action :authenticate, only: [:create, :edit]
 
   def show
     posts = Post.includes(:user, :category, :products).find(params[:id])
     render json: posts, status: :ok
+  end
+
+  def edit
+    post = @current_user.posts.includes(:products).find(params[:id])
+    render json: post, serializer: EditPostSerializer, status: :ok
   end
 
   def create
@@ -16,12 +21,12 @@ class Api::V1::PostsController < ApplicationController
         products = JSON.parse(product_params[:products])
 
         products.each do |product|
-          new_product = Product.find_or_create_by!(product_url: product['itemUrl']) do |prod|
-            prod.name = product['itemName']
-            prod.details = product['itemCaption']
-            prod.price = product['itemPrice']
-            if product['mediumImageUrls'].present?
-              uri = URI.parse(product['mediumImageUrls'])
+          new_product = Product.find_or_create_by!(product_url: product['productUrl']) do |prod|
+            prod.name = product['name']
+            prod.details = product['details']
+            prod.price = product['price']
+            if product['image'].present?
+              uri = URI.parse(product['image'])
               prod.image.attach(io: uri.open, filename: "#{prod.id}.jpg")
             end
           end
