+++ backend/app/controllers/api/v1/products_controller.rb
@@ -11,12 +11,12 @@ class Api::V1::ProductsController < ApplicationController
 
     response.each do |product|
       products << {
-        item_code: product['item_code'],
-        item_name: product['item_name'],
-        item_caption: product['item_caption'],
-        item_price: product['item_price'],
-        mediumImageUrls: product['mediumImageUrls'][0].gsub('?_ex=128x128', ''),
-        item_url: product['affiliate_url']
+        id: product['item_code'],
+        name: product['item_name'],
+        details: product['item_caption'],
+        price: product['item_price'],
+        image: product['mediumImageUrls'][0].gsub('?_ex=128x128', ''),
+        product_url: product['affiliate_url']
       }
     end
