+++ backend/app/serializers/edit_category_serializer.rb
@@ -1,0 +1,3 @@
+class EditCategorySerializer < ActiveModel::Serializer
+  attributes :id
+end
