+++ backend/config/routes.rb
@@ -6,7 +6,7 @@ Rails.application.routes.draw do
     namespace :v1 do
       post "auth/:provider/callback", to: "users#create"
       get "api_check", to: "api_check#index"
-      resources :posts, only: [:create, :show]
+      resources :posts, only: [:create, :show, :edit]
       resources :categories, only: [:index]
       resources :products, only: [:show] do
         collection do
