+++ frontend/src/app/layout.tsx
@@ -4,8 +4,8 @@ import { AppRouterCacheProvider } from '@mui/material-nextjs/v14-appRouter'
 import type { Metadata } from 'next'
 import Footer from './components/Footer'
 import Header from './components/Header'
+import { FormProvider } from '@/contexts/FormContext'
 import { NotificationProvider } from '@/contexts/NotificationContext'
-import { ProductProvider } from '@/contexts/ProductContext'
 import NextAuthProvider from '@/providers/NextAuth'
 import theme from '@/styles/theme'
 import '@/styles/destyle.css'
@@ -27,13 +27,13 @@ export default function RootLayout({
           <AppRouterCacheProvider>
             <ThemeProvider theme={theme}>
               <CssBaseline />
-              <ProductProvider>
+              <FormProvider>
                 <NotificationProvider>
                   <Header />
                   {children}
                   <Footer />
                 </NotificationProvider>
-              </ProductProvider>
+              </FormProvider>
             </ThemeProvider>
           </AppRouterCacheProvider>
         </NextAuthProvider>
