+++ frontend/src/utils/fetcher.ts
@@ -1,8 +1,12 @@
 import axios, { AxiosResponse, AxiosError } from 'axios'
 
-export const fetcher = (url: string) =>
+export const fetcher = (url: string, token?: string) =>
   axios
-    .get(url)
+    .get(url, {
+      headers: {
+        'auth-token': token,
+      },
+    })
     .then((res: AxiosResponse) => res.data)
     .catch((e: AxiosError) => {
       console.log(e)
