+++ frontend/src/utils/imageFetcher.ts
@@ -1,0 +1,11 @@
+export const imageFetcher = (url: string) => {
+  return fetch(url)
+    .then((res) => res.blob())
+    .then((blob) => {
+      return new File([blob], 'image.jpg')
+    })
+    .catch((err) => {
+      console.log(err)
+      return null
+    })
+}
